+++ src/entities/club/lib/permission.ts
@@ -1,38 +1,11 @@
+import type { ClubType } from "@/entities/club/model/club";
 import type { UserRole } from "@/entities/user/model/user";
-
-type ClubPermissionType = "MAJOR_CLUB" | "AUTONOMOUS_CLUB";
-
-interface CreateStudyPermissionParams {
-  role?: UserRole;
-}
 
 interface CreateClubPermissionParams {
   role?: UserRole;
-  clubType?: ClubPermissionType;
+  clubType?: ClubType;
   isLeader?: boolean;
   isOpeningApplicationOpen?: boolean;
-}
-
-interface CreateHomebasePermissionParams {
-  role?: UserRole;
-}
-
-function isStudyManagerRole(role?: UserRole) {
-  return (
-    role === "ADMIN" ||
-    role === "DORMITORY_MANAGER" ||
-    role === "STUDENT_COUNCIL"
-  );
-}
-
-export function createStudyPermission({ role }: CreateStudyPermissionParams) {
-  const isManager = isStudyManagerRole(role);
-
-  return {
-    isManager,
-    canManage: isManager,
-    canCheckAttendance: isManager,
-  };
 }
 
 export function createClubPermission({
@@ -44,13 +17,17 @@ export function createClubPermission({
   const isAdmin = role === "ADMIN";
   const isStudentCouncil = role === "STUDENT_COUNCIL";
   const isClubManager = isAdmin || isStudentCouncil;
+  const canManageStudy =
+    role === "ADMIN" ||
+    role === "DORMITORY_MANAGER" ||
+    role === "STUDENT_COUNCIL";
   const isMajorClub = clubType === "MAJOR_CLUB";
 
   return {
     isManager: isClubManager,
     isAdmin,
-    canCheckOpeningStatus: isStudyManagerRole(role),
-    canRegisterClub: isOpeningApplicationOpen && isStudyManagerRole(role),
+    canCheckOpeningStatus: canManageStudy,
+    canRegisterClub: isOpeningApplicationOpen && canManageStudy,
     canViewOpeningRequests: isClubManager,
     canExport: isAdmin,
     canApprove: isLeader || isAdmin,
@@ -60,13 +37,5 @@ export function createClubPermission({
     canViewApplications: isMajorClub && (isLeader || isClubManager),
     canTransferLeader: isLeader,
     canEditClub: isLeader,
-  };
-}
-
-export function createHomebasePermission({
-  role,
-}: CreateHomebasePermissionParams) {
-  return {
-    canReserve: role !== undefined,
   };
 }
