+++ src/entities/homebase/lib/permission.ts
@@ -1,0 +1,13 @@
+import type { UserRole } from "@/entities/user/model/user";
+
+interface CreateHomebasePermissionParams {
+  role?: UserRole;
+}
+
+export function createHomebasePermission({
+  role,
+}: CreateHomebasePermissionParams) {
+  return {
+    canReserve: role !== undefined,
+  };
+}
