+++ src/features/club/ui/ClubExportButton.tsx
@@ -5,7 +5,7 @@ import { useSuspenseQuery } from "@tanstack/react-query";
 import { toast } from "sonner";
 import { userQueries } from "@/entities/user/api/userQueries";
 import { exportClubs } from "@/entities/club/api/exportClubs";
-import { createClubPermission } from "@/entities/user/lib/permission";
+import { createClubPermission } from "@/entities/club/lib/permission";
 import Back from "@/shared/asset/svg/Back";
 
 export default function ClubExportButton() {
