+++ src/features/self-study/ui/SelfStudySection.tsx
@@ -8,8 +8,8 @@ import { NumberButton } from "@/shared/ui/Button/NumberButton";
 import TextField from "@/shared/ui/textField";
 import { dormitoryQueries } from "@/entities/dormitory/api/dormitoryQueries";
 import { isStudyApplicationTime } from "@/entities/dormitory/lib/applicationTime";
+import { createStudyPermission } from "@/entities/dormitory/lib/studyPermission";
 import { userQueries } from "@/entities/user/api/userQueries";
-import { createStudyPermission } from "@/entities/user/lib/permission";
 import { createApplicationActionState } from "@/shared/lib/applicationActionState";
 import { getStudyApplyButtonState } from "../lib/getStudyApplyButtonState";
 import { useCurrentTime } from "@/shared/lib/useCurrentTime";
